Performance: Optimize the AddTemplate component used in data views pages
